Inclusion of Liu and Hong scattering database.
The scattering databases provide absorption, scattering, extinction, and backscattering coefficients as well as the scattering matrix element p11 for 10 (Liu) and 6 (Hong) particles. Hong yould provide as well p22, p12, p21, p33, p44, p34, and p43. These are not yet used. Furthermore the phasefunction is trasnformed to Legendre polynomians and than to the scattering matrix with scatcnv. This detour might not be necessary.