This commit considers an issue of FASTEM and TESSEM2 with nadir observations or the emissivity and reflectivity in zenith direction.
Theoretically the emissivities and therefore the brightness temperature at nadir observations should be the same for vertical and horzontal polarization over water surfaces. In the parameterization used in FASTEM and TESSEM2 this is not the case, since they seem to be tuned for 50 and 200 GHz and have deviations at other frequencies. This bug or deficit has been reported to Catherine Prigent and NWP SAF/RTTOV. They plan a fix in future.

For now a patch has been applied in the sense, that for the nadir emissivity (nummu = 16) we take the mean of v and h for FASTEM 4/5 and TESSEM2.

Furthermore this commit reintroduces specular surfaces and sets the default sfc_refl to 'S' which should be used in general for land surfaces according to Cathrine Prigent. 'L' is still allowed for land, but an warning is thrown. 'F' for land raises an error in the python settings check.