Importer for AVAPS dropsondes (operated e.g. on HALO).
Measurement gaps will be filled via linear interpolation and missing measurements between the
aircraft altitude and the first recorded dropsonde measurement are filled with extrapolation. The
extrapolation is based on BAHAMAS measurements (if available) or otherwise guessed from the
existing measurements. After having measurement gaps filled, outliers are marked and, if a certain
threshold is exceeded, smoothed out. Finally, the dropsonde measurements are interpolated on a 10
m grid.